Fix the build of the tests (on Linux)
at least on Linux, many of the tests (TestSimpleMain, BoxTest, ...)
need the Carbon stubs.
However, they were part of MAC_SRC which was only used for the
systemshock executable itself, not the tests.
My slightly hacky "fix" is to make it part of LG_SRC instead
(because everybody links against LG_LIB anyway).