Linux: Use fopen_caseless() instead of fopen() => I can actually play!
.. except for the tests that don't open normal game assets from res/
anyway.
Maybe a reason the game worked so badly on Linux (for me) was that
res/data/objprop.dat, res/data/textprop.dat etc couldn't be opened
(in my sshock installation the path is in fullcaps for some reason)

Now I can actually progress to the second room and destroy one of the
robots and it only crashes when attacking the second robot.
Progress! \o/

Oh, also: gets() is evil, use fgets() instead in fixtest.c