Fix division by zero calculating pyramid vectors
When calculating the view pyramid vector the game tries to divide all
three components by one whose magnitude is at least 0x8000. If all three
components are below this threshold it ends up dividing by the last one,
which may be very close to zero.

To avoid a floating point exception, always divide by the longest of
the three components.