Fix drawing order of two-dimensional objects
The algorithm for sorting objects to determine draw order uses three
local variables holding the camera position, and expects them to be in
subsequent memory locations. Apparently the Mac developers ran into an
issue with their compiler placing the variables in reverse order and
"fixed" it by swapping the declaration order.

Use a portable expression to get the right coordinate instead of relying
on compiler interals. This fixes an issue with a control pedestal being
visible through an opaque bridge right at the beginning of the game when
compiling with gcc.