Improve slider operation without constraining mouse
The slider widgets in the options menu change the pointer and try to
constrain the mouse to the slider region as soon as a button is pressed.
Since mouse_constrain_xy() is currently not implemented the mouse can
leave the slider region and the widget will miss the release event.

To simplify things, just update the slider value for each button event
in the slider region, or for move events with held button.