Fix background of computer screens with text
Texture maps with text written on them were originally drawn with
background color 0x00 (black); a hack with unknown purpose changed this
to 0xFF (a shade of yellow).

We fix up the game palette to draw 0xFF as black, too, but not the
shading table. If the surface is lighted the color is still mapped to a
darker shade of yellow.