Fix animated screen not loaded crash
Crash setup:
Load a save game just before entering cyberspace to fight shodan.
Win or lose game and return to menu.
Load save game in west elevator on level 2.
Walk into the hallway near the animated screens in the room full of bodies.
As soon as the game tries to render the screens the game crashes.

What happens is that in show_obj() the animated screen texture tries to load but fails (there is a message in the console). I can't find the cause, but in the meantime prevent the crash by checking for a null bitmap. The screens on the walls won't appear, though.

Exiting the program, then loading the save and walking near the screens fixes the problem. It might be related to some quirk of how resources are handled.