Add files via upload
The CMakeLists.txt file has been edited so that Shockolate can be compiled on Raspbian Buster without needing to run ./build_deps.sh and will instead use the binaries for SDL2 and SDL2 mixer as provided by the Raspbian repository. This is a major time saver when compiling Shockolate on the Raspberry Pi. This in turn should mean that Shockolate will compile on Debian without any issues in a few weeks when Debian Buster is released. 

Here's the changes I've made:

-Removed the m32 compiler flags since Raspbian's compiler does not recognize it.
- Changed the find_library command for SDL2 to find_package.
- Removed NO_DEFAULT_PATH from the find_command for SDL2 Mixer.